requirements.txt
Flask
python-dotenv
requests
SQLAlchemy
Flask-SQLAlchemy
Flask-Login
psycopg2